autonomous_ai_system/visual_ai.py
"""
视觉推演模块 - 基于深度学习的图像理解与预测系统
支持图像识别、场景预测、视觉推理
"""
import torch
import torch.nn as nn
import torch.nn.functional as F
import torchvision.transforms as transforms
from PIL import Image
import numpy as np
from typing import Dict, List, Tuple, Optional
import cv2


class VisualAttentionModule(nn.Module):
    """视觉注意力模块"""

    def __init__(self, in_channels: int):
        super(VisualAttentionModule, self).__init__()

        self.attention = nn.Sequential(
            nn.Conv2d(in_channels, in_channels // 8, kernel_size=1),
            nn.ReLU(),
            nn.Conv2d(in_channels // 8, in_channels, kernel_size=1),
            nn.Sigmoid()
        )

    def forward(self, x):
        attention_weights = self.attention(x)
        return x * attention_weights


class VisualReasoningCell(nn.Module):
    """视觉推理单元 - 用于预测和推演"""

    def __init__(self, hidden_size: int):
        super(VisualReasoningCell, self).__init__()

        # ConvLSTM用于时空推理
        self.conv_lstm_cell = nn.Conv2d(
            hidden_size, hidden_size * 4,
            kernel_size=3, padding=1
        )

        self.attention = VisualAttentionModule(hidden_size)

    def forward(self, x, hidden_state):
        """
        Args:
            x: 当前帧特征 [B, C, H, W]
            hidden_state: 前一状态 [B, C, H, W]
        """
        if hidden_state is None:
            hidden_state = torch.zeros_like(x)

        # ConvLSTM计算
        combined = torch.cat([x, hidden_state], dim=1)
        gates = self.conv_lstm_cell(combined)

        # 分解门控
        i, f, o, g = torch.split(gates, gates.size(1) // 4, dim=1)

        i = torch.sigmoid(i)  # 输入门
        f = torch.sigmoid(f)  # 遗忘门
        o = torch.sigmoid(o)  # 输出门
        g = torch.tanh(g)     # 候选值

        # 更新状态
        new_state = f * hidden_state + i * g
        output = o * torch.tanh(new_state)

        # 应用注意力
        output = self.attention(output)

        return output, new_state


class VisualSubAI(nn.Module):
    """
    视觉子AI - 具备完整视觉推演能力

    功能:
    1. 图像识别与理解
    2. 场景预测 (基于历史帧预测未来)
    3. 视觉推理 (理解因果关系)
    4. 注意力机制
    """

    def __init__(self, config: Dict, ai_id: str):
        super(VisualSubAI, self).__init__()

        self.ai_id = ai_id
        self.config = config

        # 设备
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

        # 特征提取器 (CNN)
        self.feature_extractor = nn.Sequential(
            # 第一层
            nn.Conv2d(3, 64, kernel_size=7, stride=2, padding=3),
            nn.BatchNorm2d(64),
            nn.ReLU(),
            nn.MaxPool2d(kernel_size=3, stride=2, padding=1),

            # 第二层
            nn.Conv2d(64, 128, kernel_size=3, stride=1, padding=1),
            nn.BatchNorm2d(128),
            nn.ReLU(),
            nn.MaxPool2d(kernel_size=2, stride=2),

            # 第三层
            nn.Conv2d(128, 256, kernel_size=3, stride=1, padding=1),
            nn.BatchNorm2d(256),
            nn.ReLU(),

            # 第四层
            nn.Conv2d(256, 512, kernel_size=3, stride=1, padding=1),
            nn.BatchNorm2d(512),
            nn.ReLU(),
        ).to(self.device)

        # 注意力模块
        self.attention = VisualAttentionModule(512).to(self.device)

        # 视觉推理单元 (用于预测)
        self.reasoning_cell = VisualReasoningCell(hidden_size=512).to(self.device)

        # 预测头 (预测未来帧)
        self.prediction_head = nn.Sequential(
            nn.Conv2d(512, 256, kernel_size=3, padding=1),
            nn.ReLU(),
            nn.ConvTranspose2d(256, 128, kernel_size=4, stride=2, padding=1),
            nn.ReLU(),
            nn.ConvTranspose2d(128, 64, kernel_size=4, stride=2, padding=1),
            nn.ReLU(),
            nn.Conv2d(64, 3, kernel_size=3, padding=1),
            nn.Sigmoid()  # 输出范围 [0, 1]
        ).to(self.device)

        # 分类头 (图像识别)
        self.classifier = nn.Sequential(
            nn.AdaptiveAvgPool2d((1, 1)),
            nn.Flatten(),
            nn.Linear(512, 256),
            nn.ReLU(),
            nn.Dropout(0.5),
            nn.Linear(256, 1000)  # ImageNet类别数
        ).to(self.device)

        # 图像预处理
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406],
                               std=[0.229, 0.224, 0.225])
        ])

        print(f"[VisualSubAI {ai_id}] 初始化完成")

    def understand_image(self, image_path: str) -> Dict:
        """
        理解图像内容

        Returns:
            包含类别、置信度、注意力图的字典
        """
        # 加载图像
        image = Image.open(image_path).convert('RGB')
        image_tensor = self.transform(image).unsqueeze(0).to(self.device)

        # 提取特征
        with torch.no_grad():
            features = self.feature_extractor(image_tensor)
            attention_features = self.attention(features)

            # 分类
            logits = self.classifier(attention_features)
            probs = F.softmax(logits, dim=1)

            # 获取top-5预测
            top5_probs, top5_indices = torch.topk(probs, 5)

        results = {
            "predictions": [
                {"class": idx.item(), "confidence": prob.item()}
                for idx, prob in zip(top5_indices[0], top5_probs[0])
            ],
            "features": attention_features.cpu().numpy()
        }

        return results

    def predict_future_frames(self, video_frames: List[np.ndarray],
                            num_future_frames: int = 5) -> List[np.ndarray]:
        """
        基于历史视频帧预测未来帧

        Args:
            video_frames: 历史帧列表 [H, W, 3]
            num_future_frames: 要预测的未来帧数

        Returns:
            预测的未来帧列表
        """
        # 预处理帧
        frames_tensor = []
        for frame in video_frames:
            frame_pil = Image.fromarray(cv2.cvtColor(frame, cv2.COLOR_BGR2RGB))
            frame_tensor = self.transform(frame_pil)
            frames_tensor.append(frame_tensor)

        frames_tensor = torch.stack(frames_tensor).to(self.device)  # [T, 3, H, W]

        # 提取特征序列
        with torch.no_grad():
            features = self.feature_extractor(frames_tensor)  # [T, 512, H, W]

        # 逐步预测未来帧
        hidden_state = None
        predicted_frames = []

        for t in range(features.size(0)):
            hidden_state = features[t]

        for _ in range(num_future_frames):
            # 使用推理单元
            predicted_feature, hidden_state = self.reasoning_cell(
                torch.zeros_like(hidden_state),
                hidden_state
            )

            # 解码为图像
            predicted_image = self.prediction_head(predicted_feature)
            predicted_frames.append(predicted_image.cpu())

        return predicted_frames

    def visual_reasoning(self, image_a_path: str, image_b_path: str) -> Dict:
        """
        视觉推理 - 分析两张图像之间的关系

        Returns:
            相似度、差异分析、可能的因果关系
        """
        img_a = Image.open(image_a_path).convert('RGB')
        img_b = Image.open(image_b_path).convert('RGB')

        tensor_a = self.transform(img_a).unsqueeze(0).to(self.device)
        tensor_b = self.transform(img_b).unsqueeze(0).to(self.device)

        with torch.no_grad():
            feat_a = self.feature_extractor(tensor_a)
            feat_b = self.feature_extractor(tensor_b)

            # 计算相似度
            feat_a_flat = F.adaptive_avg_pool2d(feat_a, (1, 1)).flatten()
            feat_b_flat = F.adaptive_avg_pool2d(feat_b, (1, 1)).flatten()

            similarity = F.cosine_similarity(feat_a_flat, feat_b_flat, dim=0)

        return {
            "similarity": similarity.item(),
            "relationship": "similar" if similarity > 0.7 else "different",
            "confidence": abs(similarity.item())
        }

    def save_model(self, path: str):
        """保存模型"""
        torch.save({
            'model_state_dict': self.state_dict(),
            'config': self.config,
            'ai_id': self.ai_id
        }, path)
        print(f"[VisualSubAI {self.ai_id}] 模型已保存到 {path}")

    def load_model(self, path: str):
        """加载模型"""
        checkpoint = torch.load(path, map_location=self.device)
        self.load_state_dict(checkpoint['model_state_dict'])
        print(f"[VisualSubAI {self.ai_id}] 模型已从 {path} 加载")


class MultimodalSubAI(nn.Module):
    """
    多模态子AI - 结合视觉与语言
    """

    def __init__(self, config: Dict, ai_id: str):
        super(MultimodalSubAI, self).__init__()

        self.ai_id = ai_id
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

        # 视觉编码器
        self.visual_encoder = nn.Sequential(
            nn.Conv2d(3, 64, 7, 2, 3),
            nn.ReLU(),
            nn.AdaptiveAvgPool2d((1, 1)),
            nn.Flatten(),
            nn.Linear(64, 256)
        ).to(self.device)

        # 语言编码器
        self.text_encoder = nn.Sequential(
            nn.Linear(768, 512),
            nn.ReLU(),
            nn.Linear(512, 256)
        ).to(self.device)

        # 跨模态注意力
        self.cross_attention = nn.MultiheadAttention(
            embed_dim=256,
            num_heads=8,
            batch_first=True
        ).to(self.device)

        print(f"[MultimodalSubAI {ai_id}] 初始化完成")

    def process_vision_language(self, image: Image.Image, text: str) -> Dict:
        """处理视觉-语言联合输入"""
        # 这里需要实际的文本编码器(如BERT)
        # 简化演示

        # 图像编码
        image_tensor = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor()
        ])(image).unsqueeze(0).to(self.device)

        visual_features = self.visual_encoder(image_tensor)

        # 文本编码 (简化)
        text_features = torch.randn(1, 256).to(self.device)  # 实际应使用BERT

        # 跨模态融合
        visual_features_expanded = visual_features.unsqueeze(1)
        text_features_expanded = text_features.unsqueeze(1)

        fused_features, _ = self.cross_attention(
            visual_features_expanded,
            text_features_expanded,
            text_features_expanded
        )

        return {
            "visual_features": visual_features.cpu().numpy(),
            "fused_features": fused_features.squeeze(0).cpu().numpy()
        }


if __name__ == "__main__":
    # 测试视觉AI
    config = {
        "input_channels": 3,
        "hidden_layers": [64, 128, 256, 512],
        "kernel_size": 3,
        "use_attention": True
    }

    visual_ai = VisualSubAI(config, "test_visual_ai")
    print(f"\n视觉AI模型参数数量: {sum(p.numel() for p in visual_ai.parameters()):,}")
