autonomous_ai_system/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0
flask>=2.3.0
flask-cors>=4.0.0
